requirements.txt
ipykernel
jupyter
ipywidgets
crewai
crewai-tools 
langchain
langsmith
langchain-google-genai
python-dotenv
agentops
pytrends
pyairtable
replicate